refactor(pegaflow-transfer): split sideway backend into modules
